Change implementation to use LineList
Implementation changed the code so that `LineList_t` is used to store the
list of todo lines. This is in contrast to what was being used before,
which was an array of `Line_t`.

This had to be done so that adding and especially DELETING lines will be
easier to handle.