deletion of todo complete
- Needs further testing
- Will have to refactor some code such as replacing scrn->lines->size
with a function that tests for whether it is empty or not.
- Will have to clean up code in a major way (next thing after
implementing functionality to check a box).

- ui_refresh_delete is an implementation of the ui_refresh function that
makes it so the todo items are probably reordered on the screen after an
item is deleted. It moves everything up by 1 so that there isn't a
random blank line where the item was deleted.
- delete_todo_item is the func that is called by the main program loop
when a user presses the 'd' key in order to delete an item. A pop up
will appear asking the user to confirm.