Space toggles done/todo highlighting & item.status
I still have to change the text in the todo item to reflect the new item
status.
So the aim is for the spacebar to toggle whether the item has been
completed or not.
For example: ':TODO: Buy sand.' -> SPACEBAR -> ':DONE: Buy sand.'

makefile change

toggling todo status fin, box replaces tag in ui