current item hl function merging ++ print helper
I merged the two functions that highlight and remove the highlighting
from the item where the cursor currently is. This is because they were
not being used elsewhere and I thought it would be tidier.

Also made a new function line_wprint that will print to the Line items
window. This replaces all the times we reference line->window in order
to print something to the screen.