Helpers to get str's for enums, rm tag from line
When the file is loaded, the lines have their tags cut off. This means I
won't have something printed like this: "[x]:DONE: Buy a pizza" on the
screen. Instead it will just be "[x] Buy a pizza".
This looks a lot better.

There is also a function that now turns a status enum into a string, so
if we send it the "e_status_done" enum, it will return a string that
says ":DONE:".
Another function can return the associated checkbox, so
if we pass in a "e_status_done", we will receive a "[x]" string.